Exemplify an fs backed by a read-only s3 bucket
The view over the s3 bucket is rather minimal, where
one may only list the contents of a bucket, but not modify it
nor interact with the contents of objects in the bucket.

For the sake of simplicity the s3 bucket is queried
upon mounting and *never* refreshed afterwards, which means
that entries added to the bucket after mounting will not be
visible to the fs.